# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/games/{game_id}": {
            "get": {
                "description": "Возвращает все поля сущности Game для переданного UUID игры.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Получить информацию об игре",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игры",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные игры",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameTopicResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_gametopic.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_gametopic.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_gametopic.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут обработки запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_gametopic.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{game_id}/comments": {
            "get": {
                "description": "Возвращает упорядоченный по убыванию даты список комментариев к игре.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Получить список комментариев",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игры",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное число комментариев",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Сдвиг для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список комментариев и мета",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListCommentsResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут обработки запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет комментарий пользователя к указанной игре.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Постинг комментария",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игры",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса с полем user_id и text",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID созданного комментария",
                        "schema": {
                            "$ref": "#/definitions/handlers.AddCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Comment": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.Game": {
            "type": "object",
            "properties": {
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "$ref": "#/definitions/entity.GameRating"
                },
                "releasedate": {
                    "type": "string"
                }
            }
        },
...
        "entity.GameRating": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                },
                "gameid": {
                    "type": "string"
                },
                "ratings_count": {
                    "type": "integer"
                }
            }
        },
        "handlers.AddCommentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GameTopicResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Game"
                }
            }
        },
        "handlers.ListCommentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Comment"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/internal_controller_http_handlers_listcomments.Pagination"
                }
            }
        },
...
        "handlers.PostCommentRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_controller_http_handlers_addcomment.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "машинно-читаемый код ошибки",
                    "type": "string"
                },
                "message": {
                    "description": "человеко-читаемое сообщение",
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_addcomment.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_addcomment.APIError"
                }
            }
        },
...
        "internal_controller_http_handlers_gametopic.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_gametopic.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_gametopic.APIError"
                }
            }
        },
        "internal_controller_http_handlers_listcomments.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_listcomments.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_listcomments.APIError"
                }
            }
        },
        "internal_controller_http_handlers_listcomments.Pagination": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}